+++ apps/web/cypress/e2e/pages/safeapps.pages.js
@@ -15,6 +15,7 @@ const appMessageInput = 'input[placeholder="Message"]'
 const txBuilderUntrustedFallbackAlert = '[data-testid="untrusted-fallback-handler-alert"]'
 export const handlerInput = 'input[id="contract-field-handler"]'
 const decodedTxSummary = '[data-testid="decoded-tx-summary"]'
+export const cowFallBackHandlerTitle = 'div[title="CowSwapFallbackHandler"]'
 
 const addBtnStr = /add/i
 const noAppsStr = /no Safe Apps found/i
@@ -61,6 +62,7 @@ export const contractMethodSelector = 'input[id="contract-method-selector"]'
 export const testAddressValue2 = 'testAddressValue'
 export const testBooleanValue = 'testBooleanValue'
 export const testFallback = 'fallback'
+export const cowFallback = 'setFallbackHandler'
 export const customData = 'Custom hex data'
 export const testBooleanValue1 = '1 testBooleanValue'
 export const testBooleanValue2 = '2 testBooleanValue'
@@ -147,6 +149,8 @@ export const permissionCheckboxNames = {
   geolocation: 'Geolocation',
   fullscreen: 'Fullscreen',
 }
+
+export const cowFallbackHandler = 'sep:0x2f55e8b20D0B9FEFA187AA7d00B6Cbe563605bF5'
 
 export function verifyUntrustedHandllerWarningVisible() {
   cy.get(txBuilderUntrustedFallbackAlert).should('be.visible')
