+++ apps/web/cypress/e2e/regression/tx_details_createtx.cy.js
@@ -39,35 +39,6 @@ describe('Transaction details create tests', { defaultCommandTimeout: 30000 }, (
     safeapps.verifyUntrustedHandllerWarningVisible()
   })
 
-  it(
-    'Verify that no error for the COWSwap fallbackhandler on confirm tx screen',
-    { defaultCommandTimeout: 30000 },
-    () => {
-      cy.visit(constants.swapUrl + staticSafes.SEP_STATIC_SAFE_27)
-      const iframeSelector = `iframe[src*="${constants.swapWidget}"]`
-
-      wallet.connectSigner(signer)
-      swaps.acceptLegalDisclaimer()
-      cy.wait(4000)
-      main.getIframeBody(iframeSelector).within(() => {
-        swaps.switchToTwap()
-      })
-      swaps.unlockTwapOrders(iframeSelector)
-      main.getIframeBody(iframeSelector).within(() => {
-        swaps.selectInputCurrency(swaps.swapTokens.cow)
-        swaps.setInputValue(600)
-        swaps.selectOutputCurrency(swaps.swapTokens.dai)
-        swaps.confirmPriceImpact()
-        swaps.clickOnReviewOrderBtn()
-        swaps.placeTwapOrder()
-      })
-      createtx.clickOnContinueSignTransactionBtn()
-      createtx.clickOnAcknowledgement()
-      txs.verifyExecuteBtnIsVisible()
-      txs.verifyUntrustedHandllerWarningDoesNotExist()
-    },
-  )
-
   it('Verify that when the tx contains the action with an official 1.4.1 fallbackhandler contract there is no error', () => {
     cy.visit(`/apps/open?safe=${staticSafes.SEP_STATIC_SAFE_36}&appUrl=${encodeURIComponent(appUrl)}`)
     cy.enter(iframeSelector).then((getBody) => {
